requirements.txt
streamlit==1.31.1
pandas==1.5.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
